feat: Add startingRunningBalance to TokenConfig
Allows admin to “bias” the running balances on a chain for a particular L1 token. This would allow admin to start all running balances for a particular SpokePool for a particular L1 token at a certain value. For example, we want WETH on Optimism to start with 100 WETH running balance. This would mean that if 50 WETH of deposits occurred during a root bundle period, then the HubPool would send 50 WETH (100 - 50 = +50) to the OptimismSpokePool. Currently, the system would return 50 WETH (0 - 50 = -50) WETH from the OptimismSpokePool to the HubPool.