BREAKING CHANGE: isRootBundleValid needs to look for executed leaves up until next root bundle block (#178)
* BREAKING CHANGE: isRootBundleValid needs to look for executed leaves up until next root bundle block

Current implementation incorrectly will only look for executed pool rebalance leaves up until the end mainnet block for a bundle, which is the latest mainnet block at proposal time minus a `blockRangeEndBuffer` by design. This buffer is important throughout the dataworker, but not when validatinga root bundle. This buffer causes the dataworker to unintentionally invalidate proposed root bundles because it can miss some executed pool rebalance leaves that are executed within the `blockRangeEndBuffer`.

This ultimately has consequences on the dataworker and can cause it to have errors in computing running balances like the "slow relay excess" amount that should be subtracted from running balances.

This error occurs because `endBlockForMainet`

* fix

* Fix

* Update HubPoolClient.ts

* add test

Signed-off-by: nicholaspai <npai.nyc@gmail.com>