fix: Try disconnecting from redisClient on exception
This is currently necessary because testing has shown that redisClient
sessions can hang around indefinitely in polling mode - even if an
exception has been thrown and all context was discarded. A common Redis
configuration seems to permit up to 100 concurrent sessions, after which
it refuses any additional connections.

Per the existing implementation, an additional redisClient session will
be established whenever a bot with a non-zero POLLING_DELAY encounters
an exception. If the exception is due to a transient RPC outage, this
may result in several additional sessions being established, and given
enough time, will ultimately lead to session exhaustion and the bot will
stall indefinitely.

This workaround is definitely a bodge and it seems like it shouldn't be
necessary, but will at least help to smooth over undesirable symptoms
until a more coherent fix is available. More work is needed to
understand why the active session isn't automatically dropped once the
redisClient variable goes out of scope.

Also note that it's entirely possible that there are latent issues with
other resources here, so further cleanup may be needed.

Ref: ACX-32