fix: Dataworker should filter out all deposits younger than bundle end block
Same motivation and logic as https://github.com/across-protocol/relayer-v2/pull/221 but this filter on fillWithBlock is set higher in the function.

I don't know if this fixes the underlying bug, but the current situation is that sometimes proposer and disputer see the exact same deposits and fills in the range but construct different `netSendAmounts/runningBalances`, with the disputer consistently seeing lower values.

My guess is that there is some bug in `dataworker/PoolRebanceUtils.ts/subtractExcessFromPreviousSlowFillsFromRunningBalances()` which could only lower the net send amounts.