improve: Remove redundant processCrash() calls
This removes confusion about where the process exits in the event of an
unhandled exception with a POLLING_DELAY of 0. As a side-effect, it does
mean that the process will now unconditionally try to gracefully tear
down any existing Redis connection, which is technically redundant with
a POLLING_DELAY of 0, and *may* throw an additional exception. However,
this is fine for the improved code simplicity.

Ref: ACX-32