improve: Drop excess per-chain multicall chunks
The current implementation doesn't work as intended due to the fact that
successive chunks are submitted in parallel, so they all tend to get the
same nonce. This means that a maximum of one multicall txn can succeed
per chain. On top of this, the successful multicall seems to often be
the smallest, which means that a large number of fills are "blocked"
until demand slows down and the bot has a chance at submitting the same
multicall txn without any new Deposit candidates arriving.

Ref: ACX-98