chore: Ignore bundles on chain 1 with USDT payments (#291)
Admin bundle 357 replaces bundles 322 and 321 that cannot be executed on current Ethereum_SpokePool. 357 is executed but won't be able to be reproduced by dataworker since its construction bypasses dataworker logic. When Ethereum_SpokePool is upgraded so that 322 and 321 could be executed, we should NOT execute these or it will result in a deficit of USDT and excess payment to the relayer