improve: Ignore "missing revert data..." reason
This PR ignores the following revert reason:

> "missing revert data in call exception; Transaction reverted without a reason string"

Ignoring this reason potentially includes false positives of reverts that we should be alerted on, however there is something likely broken in how the provider is interpreting contract reverts. Currently, there are a lot of duplicate transaction sends that are reverting with this reason, for example, sending a transaction to execute a relayer refund leaf takes a while to execute and ends up reverting because a duplicate transaction mines before it. This situation leads to this revert reason which is spamming the Logger currently.

So, this PR reduces alert noise at the cost of introducing false positive ignored errors