improve: Store sped-up relayer fee % in cache
When updating deposits following speed ups, the Dataworker cached deposit does not get updated. There is no inherent risk. It does just mean when the Spokepool client updates, and loads deposits from cache, it won't be the most up-to-date version containing the new relayerFeePct.

In practice, deposits are only loaded from the cache if they are older than the most recently validated root bundle, and most deposits are fully filled well before then, but this PR will enhance UX for the edge cases where deposits only get partially filled and are not fully filled for a long time.