chore: Update dotenv 10.0.0 -> 16.0.x
This change formalises our dependency on dotenv 15.0.0, which is
needed because our .env configuration syntax requires multi-line
parsing for some configuration items, i.e.:
 - MIN_DEPOSIT_CONFIRMATIONS
 - RELAYER_INVENTORY_CONFIG

In practice, we seem to already have this version installed via some
other dependencies, as this upgraded did not result in any change to
yarn.lock.

Additionally, specify dotenv as a raw dependency, rather than a dev
dependency, since non-developers are reliant on it.