refactor: Remove dataworker deposit/fill event caching logic
This reverts most of the code originally introduced in #192 that enabled the `SpokePoolClient` to cache `Deposit` and `Fill` events. Note that this PR does not touch the other use of the cache, to fetch `realizedLpFeePct` computations from the cache.

The reason is that this feature is no longer used because even with the cache, the time to load all events from the beginning of time for Across is too large. Therefore, the Dataworker has shifted towards an optimization strategy that fixes its lookback window rather than caches all events.

A followup PR could introduce a further optimization that could cache the data loaded from the fixed lookback. This could work because during the proposal challenge period, a lot of the events loaded from the fixed dataworker lookback are duplicates, particularly those that are older than the pending bundle block range.

The main benefit of this PR is to remove code