improve: Use Deposit/Fill event cache with fixed lookback for Dataworker
Will improve average time of dataworker. Before a new bundle is proposed, or the latest one is fully executed, the Dataworker uses the same lookback (i.e. the same `fromBlock` to look to). This means that the dataworker is making a lot of duplicate queries to an RPC, which can be cached.

This PR means that each time a new bundle is fully executed, the cache will be updated and new events will be loaded, but until the subsequent bundle is executed, the cache should offer some speed up