improve(ProfitClient): Remove ignoreTokenPriceFailures option
This variable seems pretty redundant now. Price lookup failures are very
rare these days because we have fallback from the Across API to
CoinGecko Free/Pro, and DefiLlama exists as a third option if we choose
to use it. Additionally, the ignoreProfitability option is another good
proxy for whether we want to be strict about price lookup failures.