improve(MultiCallerClient): Factor out txn simulation
Reduce the size/complexity of the executeTransactionQueue() method by
factoring out the transaction simulation component. There is no
functional impact as part of this change - the relocated code performs
the same function as before.

Follow-up commits might make tweaks to improve transaction simulation
now that it is confined to a separate private method.