fix(MultiCallerClient): Correct multicall txn validation
The existing implementation incorrectly checks that _each_ enqueued
transaction has the wrong contract address. This is impossible because
the contract address used as a reference is sourced from the first
transaction in the queue.

While here, add a test for this and expand the logic to validate that
all enqueued transactions are destined for the same chain ID as well..