improve(MultiCallerClient): Simplify sim failure logging (#418)
Makes the code a bit more compact and more clearly separates between
simulation and logging. This also allows for the logging function to
be verified in test by mocking it out.

As a minor benefit, the existing implementation performs 3 complete
loops over the array of transaction simulation results. This change
performs only one complete loop, and then a secondary loop over the
list of failures to determine which to log. It should be marginally
faster, and probably more memory efficient.