fix(MultiCallerClient): Correct multicall txn validation (#417)
The existing implementation incorrectly checks that _each_ enqueued
transaction has the wrong contract address. This is impossible because
the contract address used as a reference is sourced from the first
transaction in the queue, so there will always be _at least_ one enqueued
transaction that fails this test. This effectively means that the subsequent
code block is dead code.

While here, add a test for this and expand the logic to validate that
all enqueued transactions are destined for the same chain ID as well.