improve(MultiCallerClient): Use txnClient for simulation
This is _mostly_ done now because it significantly reduces the
complexity of the subsequent diff that introduces the more
significant MultiCallerClient changes.

I had to disable a single test as part of this change. This was done so
as to avoid making short-term changes to the existing MultiCallerClient
implementation: modifying that code too much is an unnecessary risk for
no real benefit.