chore(config): Prevent non-zero POLLING_DELAY
This is an administrative change to temporarily prevent operators from
running in looping mode. The justification for this is that we
currently do not apply the same QA to non-zero polling delay (looping
mode) as we do to serverless mode.

Looping mode is useful for third-party operators and the intention is
ultimately to support it.