chore(ProfitClient): Remove ignoreProfitability (#457)
ignoreProfitability has not been used since late October, and it was
marked in the code as a temporary feature. Now that we have minimum
fee granularity at the per token/route level, it seems like can probably do
without this config item.

For reference, it should still be possible to emulate ignoreProfitability
by setting GAS_MULTIPLIER=0.01 to ignore the cost of a making fills,
and then and tuning any token/route-specific minimums as necesary.