feat(Monitor): Add Refiller
Refiller is a new Monitor mode that will refill balances. The user must set a trigger and target per account, chain, token combination. The caller is responsible for setting these triggers above the monitored targets/thresholds so as not to create a lot of alert noise. I think this is a reasonable ask of the bot runner since the refiller is sending transactions on-chain. Moreover, the refiller will likely only be refilling a couple accounts.

The initial accounts that should be added are the dataworker and hubpool addresses, as the relayer already has its own inventory management logic.