feat(uba): Support SpokePoolClient test event injection
This change implements a mocked SpokePoolClient, with the
SpokePoolClient._update() method overridden in order to suppress
on-chain queries and instead inject ethers Event objects directly.

It _should_ have a noticeable speed difference over running end-to-end
test via Hardhat, and it permits test of Events that are not yet emitted
by the contracts-v2 package. With this feature we can develop further
without a hard dependency on bumping the contracts ABI. Further, it
should be trivially reusable for other SpokePoolClient events in test.