improve(SpokePoolClient): Optimise EnabledDepositRoute
Per current logic, we must search for EnabledDepositRoute events back
to SpokePool deployment. Whether or not this is desirable for all bot
instances, we can optimise it by dramatically increasing the block
range over which we are searching. This should result in fewer RPC
eth_getLogs() queries for this topic.