fix: :bug: improve logging support for failed multicall txns
This change adds a logging and exit condition to the multicaller txn execution. In the event that the transactions fail, the new logic will collate failed txn states and produce an error log. Additionall this fn will propagate the error and lead to an eventual (caught) stack unwind to prevent any further logic