refactor(SpokePoolClient): Split update() and add validation (#523)
Split update(), such that the parts specifically querying the RPC are performed in
a new method named _update(). This significantly reduces the length of update()
and hopefully makes it easier to reason about when it's OK to conclude that
SpokePoolClient updating was successful.

In addition:
 - Be more strict when validating the responses to the following calls:
    - spokePool.numberOfDeposits()
    - spokePool.getCurrentTime()
    - provider.getBlockNumber()
 - Generate array of known event names once in the constructor, not on each call
    to update().

As a side benefit, because _update() is split off into a separate method, it's now easy
to mock it out in test and inject a variety of data to verify the remaining update()
behaviour. That capability will be added in a follow-up commit.