fix(SpokePoolClient): Don't reorder on getDeposits()
getDeposits() internally sorts deposit events according to blockNumber.
The deposit blockNumber is however the block number corresponding to the
quoteTimestamp, not the block number of the deposit transaction. This is
a quirk that seems unique to FundsDeposited events.

This method is only used in test, but the re-ordering is incompatible
with the new UBA design, where the ordering of SpokePool transactions
takes precedence.