improve(FundsDeposited): Align blockNumber purpose
The FundsDeposited event is unique in that its blockNumber currently
refers to the block number according to its quoteTimestamp. All other
events use blockNumber to specify the timestamp of their event.

This change re-aligns FundsDeposited to be more compatible with the
other type definitions, which will make it easier to sort them for the
UBA model.