improve(UBA): Support retrieiving SpokePool flows
This change adds a helper function to retrieve and sort UBA flows.
The final resting place for this functionality is tbd, but for now it
can safely live as a utility function in utils. The primary user of
this function is expected to be the UBAFeeCalculator class, also in
utils.

The implementation implicitly (unfortunately) involves multiple rounds
of sorting, by virtue of the fact that the SpokePoolClient already
performs in-place sorting on all retrieved events before they are
parsed and stored in the SpokePoolClient. This is maintained in order
to avoid changing existing SpokePoolClient behaviour, but may be
re-evaluated as the UBA implementation matures.

Some additional modifications were needed to the mocked SpokePoolClient
implementation. These changes should incidentally make it easier to add
further tests in the future.

Fixes: ACX-1004