improve(MulticallerClient): Remove extra simulation
This PR removes a simulation of the batch sent to `Multicall3`. This was originally included because technically its possible that individually transactions will succeed when batched together, they can revert.

This extra simulation has downsides:
- Extra RPC request to simulate transaction ends up leading to slow down in transactions being mined which opens up to ultimate failures like nonce-reuse and front-running
- More logic

With some upsides:
- Catches case where batch fails while individual txns succeed

Ultimately, the dataworker and relayer functions using this code should never be batching together transactions that would fail in a batch but succeed individually, so I think its safe to remove to mitigate the downsides above. Moreover, if the batch fails, we'll hear about it anyways.

The motivation for this PR was an error log that showed the batch failed simulation, which suggests that the individual txn suceeded, but the batch failed. This error did not persist, meaning that the `fillRelay` reverted after succeeding the individual simulation but was probably front-run by the time of the batch txn.