improve(ProviderUtils): New RPC provider config syntax
This commit introduces new syntax for RPC provider configuration. With
this change, providers can be configured independently with the
following env var:

  RPC_PROVIDER_<PROVIDER_NAME>_<CHAIN_ID>=<ENDPOINT_URL>

The use and preference of a provider is determined by the ordering of
its appearance a comma-separated string in the RPC_PROVIDERS_<CHAIN_ID>

  RPC_PROVIDERS_1=<PROVIDER_1>,<PROVIDER_2>,<PROVIDER_3>,...

The new config parser will optimistically try for the new syntax, and if
it fails, will silently fall back to the existing NODE_URL_X and
NODE_URLS_X syntax. This allows users the opportunity for a seamless
migration. Ultimately, I'd propose to deprecate the existing NODE_URL*
syntax after one or more weeks, to allow sufficient time for third-party
operators to adjust their config.

Provider configuration examples:

  RPC_PROVIDER_INFURA_1=https://mainnet.infura.io/v3/...
  RPC_PROVIDER_ALCHEMY_1=https://eth-mainnet.g.alchemy.com/v2/...
  RPC_PROVIDER_QUICKNODE_1=...

  RPC_PROVIDERS_1=INFURA,ALCHEMY,QUICKNODE

To re-order, such that Alchemy is the preferred provider:

  RPC_PROVIDERS_1=ALCHEMY,INFURA,QUICKNODE

Some advantages of this new approach are:
 - The same syntax is used for when there are one or multiple RPC
   providers. The existing configuration can be unintuitive because both
   NODE_URL_X and NODE_URLS_X is supported, and this can confuse users.

 - No JSON is needed, so JSON parsing errors can not trigger a failure.
   This is especially convenient for users who manage their environemnt
   via .env.

 - It's OK to leave the list of defined providers uncommented in the
   environment, since the RPC_PROVIDERS_<CHAIN_ID> env var is actually
   what dictates which providers are used. This makes testing easier.

 - When multiple providers are used, re-ordering the providers does not
   require touching the endpoint URLs, which often long in length and
   cumbersome to adjust.

 - Per-provider configuration is an obvious follow-on (i.e. permitting
   a unique block range look back to be specified for each provider,
   rather than a global config per chain.