feat(UBA): Support overloaded ExecutedRootBundle runningBalances
The UBA needs to keep track of the size of the incentive pool for each
token on each chain. Per the pending updates to UMIP-157, this will be
done by appending the running incentive balances to the
ExecutedRootBundle runningBalances array.

This change simply updates existing tests to demonstrate that the data
can be read out and made available to callers.