refactor(ConfigStoreClient): Remove getUtilization override (#688)
This method seems surprisingly unused by any other code in this repository!

It's proposed for removal because it introduces a dependency on the
HubPoolClient, which seems like an inversion. Per a pending change to
the HubPoolClient in the sdk-v2 repository, the HubPoolClient will instead
become dependent on the ConfigStoreClient when validating the
runningBalances member of the ExecutedRootBundle.