improve(relayer): Retain deposit version numbers (#730)
Upgrade the existing "supported version" filtering logic by retaining
the applicable version number in the [Relayer]UnfilledDeposit object.
The benefit of retaining the version number is that it allows downstream 
recipients of the [Relayer]UnfilledDeposit type to implement different
behaviour based on the version number. The new method of pulling out the
version numbers should be a tiny bit more efficient too, since there's
less looping involved.

As part of this I also found that we had two competing definitions for
an UnfilledRelay - one in the SpokePool interfaces from the SDK, and
another locally defined in FillUtils. For now I have renamed them to
avoid confusion, but they'll need to be merged back in the SDK as part
of getting the dataworker ready for the UBA model.