refactor(ProfitClient): Misc improvements (#734)
Reuse fixed-point division. This saves creating a bunch of throwaway
BigNumber types just for the purpose of dividing. Additionally,
normalise log formatting of numbers back from 18 decimals. This
makes it much easier to read the messages and debug misbehaviour.

These changes are made in preparation for an update for UBA, but
were diff-noisy, so I split them out.