improve: Allow transactions to have padded gas limits
This change implements two key things:

1) Capture the gas estimates from successful txn simulations and pass it
through to txn submission.

2) Modify runTransaction() to accept a gasLimitMultiplier argument,
which will be applied if the gasLimit is also supplied.

3) Modify the MultiCallerClient class to apply 20% gas padding on
multicall3 aggregate() transactions. This addresses a specific issue
that has been seen with the executor personality of the dataworker
recently, resulting in a significant number of OoG failures.

Side benefit: This might actually save a separate estimateGas() RPC
call, so it's possibly faster and a bit cheaper in terms of RPC usage.