improve(runTransaction): Pad gas for multicall3 aggregate (#744)
This change implements three key things:

1) Capture the gas estimates from successful txn simulations and pass it
through to txn submission.

2) Modify TransactionClient() to check for a gasLimitMultiplier argument
If a multiplier was supplied, apply it to the gasLimit at that point,
before passing a transaction request over to runTransaction() for
submission to the RPC provider.

3) Modify the MultiCallerClient class to apply 50% gas padding on
multicall3 aggregate() transactions. This addresses a specific issue
that has been seen with the executor personality of the dataworker
recently, resulting in a significant number of OoG failures.

Side benefit: This might actually save a separate estimateGas() RPC
call, so it's possibly faster and a bit cheaper in terms of RPC usage.

Fixes ACX-1236.