refactor(Relayer): Remove exception wrapper in fillRelay() (#751)
This large exception wrapper is a hangover from the early days of the
relayer codebase. It's difficult to see what the encapsulated code would
be doing to generate a transient error that we'd want to suppress,
rather than fix. Additionally, I was unable to find any instances of the  
"Error creating fillRelayTx" message in the relayer logs over the past
180 days.

While here, since there are now no async calls in fillRelay(), remove
its async declaration.