improve(ProfitClient): Permit gas costs to be ignored (#752)
Recent testing has been difficult because third-party relayers are eager
to make fills, and are sometimes doing so at a loss. This means it can 
be surprisingly difficult to set a relayer fee low enough to be ignored
by others. This change permits a 0x gas multiplier to be used, thereby
instructing the relayer to ignore the cost of gas entirely. Deposits can 
thus be set with a 0% relayerFeePct, which is good enough to be ignored
by most third parties. The pre-existing profitability protections are 
still in place, and control remains in the hands of the user as to how 
they want to configure their relayer.