fix(AdapterManager): Pad gas for deposits to Optimism (#754)
As per the recent change in the MultiCallerClient, when submitting
transactions destined for the Optimism deposit contract on Ethereum,
it's necessary to pad the gas that is estimated via estimateGas(). The 
TransactionClient class has been used for this because it's a drop-in
replacement for runTransaction(), and the MultiCallerClient offers
limited benefits in this case because coincidental rebalancing to
multiple chains seems unlikely.