improve(relayer): Simplify deposit version filtering (#841)
The version-filtered array of unfilled deposits was not subsequently
used when deriving the confirmedUnfilledDeposits array that is
ultimately used to make fills. This was fortunately not a showstopper
for the relayer because we still had an old check in
`getUnfilledDeposits()`. This PR removes that and implements a test for
specific version-based filtering.

In order to make it harder to trip over this in future, factor out the
version determination so that there are fewer overall variables to track
in the checkForUnfilledDepositsAndFill() method. This also makes it easy
to add a specific unit test.