refactor(finalizer): Split out per-chain finalization (#851)
In preparation for supporting a zkSync finalizer (...and subsequently
other variants for each new chain), slice up the existing finalize()
function to make it more modular and maintainable. There's no intended
change in functionality as a result of this.

This is only a first step; there will be more opportunity for 
refactoring in follow-up work.