fix: Ignore block hash in eth_getLogs responses
We don't _directly_ depend on the block hash, so if all of the other
fields in the response agree then don't discard the response simply
because the of the blockHash field. This is introduced now because there
are upstream disagreements from various providers as to which blockHash
applies to the current set of events.