feat: Add protection against computing LP fees for future quote timestamps
This resolves the [reverted SDK PR](https://github.com/across-protocol/sdk-v2/pull/399) and re-introduces protection from deposit quote timestamps that are impossible to compute.

This PR essentially moves that protection to the production version of Hub and Spoke clients so that the unit tests can continue to work and not have to add additional logic to set quote times correctly. Ideally in the future we set the quote times how they would in production but for now this solution results in a much smaller code diff