improve: Simulate finalization txns before failure
Simulate all finalizations before bundling them into the multicall
aggregate() transaction, which will fail in aggregate if even a single
transaction is unsuccessful. This caused problems today, when a small
number of zkSync withdrawals were unsuccessful, resulting in a finalizer
halt for a few hours.

Transactions are simulated first for logging purposes; it would be
possible to user multicall3.tryAggregate(false, ...), except it's
non-trivial to extract the results of each transaction.

Fixes ACX-1551