feat: Basic support for relayer auctions
AUCTION_API_ENDPOINT must be set to the HTTP URL hosting the auction
API.

Note: There are some sharp edges for the depositor here. The depositor
should ideally be able to verify that the message and recipient proposed
by the auction API will actually result in the funds being delivered to
them.