fix(finalizer): Handle 1:many withdrawal:calldata
The previous change to the finalizer incorrectly assumed that there
would be a 1:1 relationship between a txn and a withdrawal or proof.
This is incorrect on Polygon, where 2 txns are needed to finalize a
single withdrawal.

This change seems like a manageable way of establishing a 1:many
relationship between withdrawal/proof logging and the necessary
transaction(s), without imposing any changes on the underlying chain
adapters. Ultimately it would be nice to backport this into the adapters
themselves, but that's invasive and it's not worth prioritising now.