improve: Retry Across API /limits on failure (#971)
AcrossAPIClient.update() currently only tries to query the first
available route for a token. In instances where the RPC provider for 
this chain is unavailable, this results in the client resolving a
maxDeposit of 0.

Instead, iterate over the set of all available chains for this token
until one route results in a usable response. This should boost
robustness in the bots.

While here, tweak the instantiation of the AcrossAPIClient. The existing
implementation requires RELAYER_DESTINATION_CHAINS to be defined. Per 
this PR, if RELAYER_DESTINATION_CHAINS is not defined, just use the full
set of available SpokePoolClient instances.