fix(finalizer): Revert Polygon txn:withdrawal expectations
In the case where multiple Polygon withdrawals for a single token have
accrued, the existing logic can break down. This is because the bridge
adapter retrieve() call is a catch-all that evacuates the entire
balance, to the HubPool, so multiple finalisation transactions can map
to a single retrieve() call. This is gas-efficient, but makes it
borderline impossible for the finalizer to expect a clean 1:n
withdrawal:txn mapping.