improve(relayer): Add filter for deposits whose quote blocks are in the future
The `SpokePoolClient` already filters out deposits with quote timestamps that are in the future relative to the spoke pool chain [here](https://github.com/across-protocol/sdk-v2/blob/master/src/clients/SpokePoolClient.ts#L614).

This adds further protection that the resolved mainnet block for the `quoteTimestamp` is less than the `HEAD` block on mainnet.